move feature selection code from Scene to own class
